requirements.txt
schedule
python-vlc
streamlit
